feat: 重写了 cqbot 适配器
